+++ packages/process/lib/Polysemy/Process/Data/ProcessOutputParseResult.hs
@@ -1,0 +1,8 @@
+module Polysemy.Process.Data.ProcessOutputParseResult where
+
+data ProcessOutputParseResult a =
+  Done { value :: a, leftover :: ByteString }
+  |
+  Partial { continue :: ByteString -> ProcessOutputParseResult a }
+  |
+  Fail { error :: Text }